% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryVisualiseR.R
\name{binary_visualiseR}
\alias{binary_visualiseR}
\title{Binary Visualiser - A Binary Confusion Matrix Visual}
\usage{
binary_visualiseR(
  cm_input,
  class_label1 = "Class Negative",
  class_label2 = "Class Positive",
  quadrant_col1 = "#3F97D0",
  quadrant_col2 = "#F7AD50",
  custom_title = "Confusion matrix",
  text_col = "black",
  round_dig = 2
)
}
\arguments{
\item{cm_input}{confusion matrix object to inpt}

\item{class_label1}{classification label 1 i.e. readmission into hospital}

\item{class_label2}{classification label 2 i.e. not a readmission into hospital}

\item{quadrant_col1}{colour of the first quadrant - specified as hexadecimal}

\item{quadrant_col2}{colour of the second quadrant - specified as hexadecimal}

\item{custom_title}{title of the confusion matrix plot}

\item{text_col}{the colour of the text}

\item{round_dig}{rounding options}
}
\value{
returns a data frame containing the relevant confusion matrix statistics
}
\description{
a confusion matrix object for binary classification machine learning problems.
Returns the outputs of the matrix into a row structure for storage in a database or data frame.
}
\examples{
multi_class_cm(cm_input, class_label1="Class Negative",class_label2="Class Positive")
}
\keyword{internal}
