% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_impeR.R
\name{var_impeR}
\alias{var_impeR}
\title{Variable importance - var_impeR}
\usage{
var_impeR(model)
}
\arguments{
\item{model}{a trained caret model you want to create a variable importance summary of}
}
\value{
A list with tibble and ggplot object inside
\itemize{
  \item rowid - an ordered row index
  \item Metric - the variable importance metric
  \item Overall - the variable importance weight relating to mean decrease in accuracy
  \item Prop - the proportion of the variable importance weight over all importance metrics
  \item Generated - when the variable importance plot was run
}
}
\description{
takes the outputs of the caret package to create variable importance plots.
}
\examples{
var_impeR(model) #Trained ML model from caret library
}
\references{
ggplot2
}
