% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleFramer.R
\name{binary_class_cm}
\alias{binary_class_cm}
\title{Binary Confusion Matrix data frame}
\usage{
binary_class_cm(caret_cm)
}
\arguments{
\item{caret_cm}{confusion matrix generated from the caret library}

\item{.script_name}{name to assign to the data frame for the model}
}
\value{
returns a data frame containing the relevant confusion matrix statistics
}
\description{
a confusion matrix object for binary classification machine learning problems.
}
\examples{
binary_class_cm(caret_cm_object)
}
\keyword{internal}
