% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiFramer.R
\name{multi_class_cm}
\alias{multi_class_cm}
\title{Multiple Confusion Matrix data frame}
\usage{
multi_class_cm(caret_cm)
}
\arguments{
\item{caret_cm}{confusion matrix generated from the caret library}
}
\value{
returns a data frame containing the relevant confusion matrix statistics
}
\description{
a confusion matrix object for multi classification machine learning problems.
Returns the outputs of the matrix into a row structure for storage in a database or data frame.
}
\examples{
multi_class_cm(caret_cm)
}
\keyword{internal}
